+++ frost-core/src/tests/vectors.rs
@@ -35,7 +35,7 @@ pub fn parse_test_vectors<C: Ciphersuite>(json_vectors: &Value) -> TestVectors<C
     let secret_key_str = inputs["group_secret_key"].as_str().unwrap();
     let secret_key_bytes = hex::decode(secret_key_str).unwrap();
     let secret_key =
-        SigningKey::from_bytes(secret_key_bytes.try_into().debugless_unwrap()).unwrap();
+        SigningKey::deserialize(secret_key_bytes.try_into().debugless_unwrap()).unwrap();
 
     let message = inputs["message"].as_str().unwrap();
     let message_bytes = hex::decode(message).unwrap();
@@ -330,7 +330,7 @@ pub fn check_sign_with_test_vectors<C: Ciphersuite>(json_vectors: &Value) {
 
     // Check that the generated signature matches the test vector signature
     let group_signature = group_signature_result.unwrap();
-    assert_eq!(group_signature.to_bytes().as_ref(), signature_bytes);
+    assert_eq!(group_signature.serialize().as_ref(), signature_bytes);
 
     // Aggregate the FROST signature from our signature shares
     let group_signature_result =
@@ -341,5 +341,5 @@ pub fn check_sign_with_test_vectors<C: Ciphersuite>(json_vectors: &Value) {
 
     // Check that the generated signature matches the test vector signature
     let group_signature = group_signature_result.unwrap();
-    assert_eq!(group_signature.to_bytes().as_ref(), signature_bytes);
+    assert_eq!(group_signature.serialize().as_ref(), signature_bytes);
 }